# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/requests/{id}": {
      "get": {
        "operationId": "RequestsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Requests"
        ]
      },
      "patch": {
        "operationId": "RequestsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: ADMINISTRATOR, OFFICER",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Requests"
        ]
      },
      "delete": {
        "operationId": "RequestsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: ADMINISTRATOR",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Requests"
        ]
      }
    },
...
    "/resource-requests": {
      "get": {
        "operationId": "ResourceRequestsController_findAll",
        "parameters": [
          {
            "name": "engineerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ResourceRequests"
        ]
      },
      "post": {
        "operationId": "ResourceRequestsController_create",
        "parameters": [
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: ENGINEER",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ResourceRequests"
        ]
      }
    },
    "/resource-requests/{id}": {
      "get": {
        "operationId": "ResourceRequestsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ResourceRequests"
        ]
      },
      "patch": {
        "operationId": "ResourceRequestsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: ENGINEER, OFFICER, ADMINISTRATOR",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourceRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ResourceRequests"
        ]
      },
      "delete": {
        "operationId": "ResourceRequestsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: ADMINISTRATOR, OFFICER, ENGINEER",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ResourceRequests"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "InfraLynx API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpdateRequestDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "RECEIVED",
              "UNDER_REVIEW",
              "APPROVED_FOR_PLANNING",
              "CONVERTED_TO_WORK_ORDER",
              "CLOSED",
              "REJECTED"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "urgency": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "EMERGENCY"
            ]
          }
        }
      },
...
      "CreateResourceRequestDto": {
        "type": "object",
        "properties": {
          "departmentId": {
            "type": "string"
          },
          "engineerId": {
            "type": "string"
          },
          "item": {
            "type": "string"
          },
          "quantity": {
            "type": "string"
          },
          "urgency": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "departmentId",
          "engineerId",
          "item",
          "quantity",
          "urgency",
          "status"
        ]
      },
      "UpdateResourceRequestDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "urgency": {
            "type": "string"
          }
        }
      },
...
    }
  }
}